requirements.txt
# Core ML/AI libraries
numpy==1.24.3
tensorflow==2.13.0
scikit-learn==1.3.0
scipy==1.11.1

# Spotify API
spotipy==2.23.0

# Web framework
flask==2.3.2
flask-cors==4.0.0

# Data processing
pandas==2.0.3
requests==2.31.0

# Utilities
tqdm==4.65.0
python-dotenv==1.0.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Development
jupyter==1.0.0 